# Speeds up the linear system solve in KanesMethod.rhs()