Fixes #10964.

```
moorepants@garuda:sympy(faster-kane-rhs)$ ipython
Python 3.5.1 |Continuum Analytics, Inc.| (default, Dec  7 2015, 11:16:01) 
Type "copyright", "credits" or "license" for more information.

IPython 4.1.2 -- An enhanced Interactive Python.
?         -> Introduction and overview of IPython's features.
%quickref -> Quick reference.
help      -> Python's own help system.
object?   -> Details about 'object', use 'object??' for extra details.

In [1]: from pydy.models import n_link_pendulum_on_cart

In [4]: sys = n_link_pendulum_on_cart(18)

In [5]: %time rhs = sys.eom_method.rhs()
CPU times: user 3min 36s, sys: 4 ms, total: 3min 36s
Wall time: 3min 36s

In [6]: 
Do you really want to exit ([y]/n)? 
moorepants@garuda:sympy(faster-kane-rhs)$ git checkout master
Switched to branch 'master'
Your branch is up-to-date with 'origin/master'.
moorepants@garuda:sympy(master)$ ipython
Python 3.5.1 |Continuum Analytics, Inc.| (default, Dec  7 2015, 11:16:01) 
Type "copyright", "credits" or "license" for more information.

IPython 4.1.2 -- An enhanced Interactive Python.
?         -> Introduction and overview of IPython's features.
%quickref -> Quick reference.
help      -> Python's own help system.
object?   -> Details about 'object', use 'object??' for extra details.

In [2]: from pydy.models import n_link_pendulum_on_cart

In [3]: sys = n_link_pendulum_on_cart(18)

In [4]: %time rhs = sys.eom_method.rhs()
CPU times: user 4min 17s, sys: 4 ms, total: 4min 17s
Wall time: 4min 17s
```