@moorepants 2016-04-16
@sympy/mechanics This needs a review.