@jbm950 2016-07-22
So this version took off ~40 seconds of run time? Looks good to me. I think your docstring additions in #11183 are more thorough though.

I agree it's probably best to not invert the part of the matrix that is already solved.